build(circular-deps): tsPreCompilationDeps only fix for types
