+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export { IChannel } from './channel'
-export { default as IGif, IImage, IImages, ImageAllTypes, IRendition } from './gif'
+export { default as IGif } from './gif'
+export { IImage, IImages, ImageAllTypes, IRendition } from './images'
 export { PingbackEventType } from './pingback-event'
 export { default as IUser, IProfileUser } from './user'
